Refactor compare_releases script to use clikt (#2801)
